+++ src/components/tile.tsx
@@ -1,18 +1,14 @@
 import { ITile } from '../services/types';
 
-type myProps = {
-  tile: ITile;
-};
-
-function Tile(p: myProps) {
+function Tile({ tile }: { tile: ITile }) {
   return (
     <div className="tile">
-      <img className="tile_img" src={URL.createObjectURL(p.tile.file[0])}></img>
-      <h3>{p.tile.name}</h3>
-      <p>Gender: {p.tile.gender}</p>
-      <p>Age: {p.tile.age} years</p>
-      <p>Country: {p.tile.country}</p>
-      <p>{p.tile.email}</p>
+      <img className="tile_img" src={URL.createObjectURL(tile.file[0])} />
+      <h3>{tile.name}</h3>
+      <p>Gender: {tile.gender}</p>
+      <p>Age: {tile.age} years</p>
+      <p>Country: {tile.country}</p>
+      <p>{tile.email}</p>
     </div>
   );
 }
