+++ src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import App from './App.tsx';
 import './index.css';
 import { BrowserRouter } from 'react-router-dom';
-import { setupStore } from './store/store.tsx';
+import { setupStore } from './services/store.tsx';
 import { Provider } from 'react-redux';
 
 const store = setupStore();
