--- src/store/hookReducer.tsx
@@ -1,18 +1,0 @@
-import { PayloadAction, createSlice } from '@reduxjs/toolkit';
-import { IData, IForm } from '../services/types';
-
-const defaultState: IForm = {
-  data: null,
-};
-
-export const hookSlice = createSlice({
-  name: 'tiles',
-  initialState: defaultState,
-  reducers: {
-    hookSubmit(state, action: PayloadAction<IData>) {
-      state.data = action.payload;
-    },
-  },
-});
-
-export default hookSlice.reducer;
